papers/FCENet/requirements.txt
asttokens==2.0.5
astunparse==1.6.3
certifi==2020.12.5
cffi==1.14.6
cycler==0.10.0
Cython==0.29.24
decorator==5.1.0
easydict==1.9
joblib==1.0.1
kiwisolver==1.3.2
matplotlib==3.4.3
mindspore @ https://ms-release.obs.cn-north-4.myhuaweicloud.com/1.3.0/MindSpore/cpu/x86_64/mindspore-1.3.0-cp37-cp37m-linux_x86_64.whl
mindspore-gpu @ https://ms-release.obs.cn-north-4.myhuaweicloud.com/1.3.0/MindSpore/gpu/x86_64/cuda-10.1/mindspore_gpu-1.3.0-cp37-cp37m-linux_x86_64.whl
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
numexpr==2.7.3
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603479632437/work
olefile==0.46
opencv-python==4.5.3.56
packaging==21.0
pandas==1.3.3
Pillow @ file:///tmp/build/80754af9/pillow_1615057347459/work
Polygon3==3.0.9.1
protobuf==3.18.0
psutil==5.8.0
pycocotools==2.0.2
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
scikit-learn==1.0
scipy==1.5.4
Shapely==1.8.1.post1
six @ file:///tmp/build/80754af9/six_1605205313296/work
sklearn==0.0
tables==3.6.1
threadpoolctl==2.2.0
tqdm==4.62.3
typing-extensions @ file:///home/ktietz/src/ci_mi/typing_extensions_1612808209620/work
